requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
requests==2.22.0
RPi.GPIO==0.7.0
spidev==3.4
urllib3==1.25.3
